Explicit std::min/max template usage
This is unfortunately required in several cases:
 - Sometimes when we use either float and double (this should be changed)
 - Because KDCoordinate is not an int, so any arithmemtic promotes it to
an int
 - Because we mix pointer differences and ints

+++ apps/sequence/graph/graph_controller.cpp
@@ -52,7 +52,7 @@ float GraphController::interestingXHalfRange() const {
     Sequence * s = functionStore()->modelForRecord(functionStore()->activeRecordAtIndex(i));
     int firstInterestingIndex = s->initialRank();
     nmin = std::min(nmin, firstInterestingIndex);
-    nmax = std::max(nmax, firstInterestingIndex + standardRange);
+    nmax = std::max(nmax, firstInterestingIndex + static_cast<int>(standardRange));
   }
   assert(nmax - nmin >= standardRange);
   return nmax - nmin;
